+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Semantic versioning
 Version latest (0.12.0rc?, 2019-09-24?)
 
 - ...
+
+
+Version 0.12.2 (2019-10-25)
+
 - enh: use click's exception handling for IsitfitCliError
 	- pass `is_outdated` in click context for error verbosity about upgrade
 - enh: use click's Command to display footer after invoke